carto_app/requirements.txt
Flask
folium
qrcode
pdfkit
python-docx
Pillow
